mp-linux.cpp: warning.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@26254 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r3/linux/mp-linux.cpp
@@ -120,7 +120,7 @@ static uint32_t rtMpLinuxGetFrequency(RTCPUID idCpu)
 
 RTDECL(int) RTMpCpuIdToSetIndex(RTCPUID idCpu)
 {
-    return idCpu < rtMpLinuxMaxCpus() ? idCpu : -1;
+    return idCpu < rtMpLinuxMaxCpus() ? (int)idCpu : -1;
 }
